requirements.txt
PySide6>=6.4,<6.5
Pillow>=10.0
numpy>=1.25,<2
torch>=2.0
